requirements.txt
flask>=2.0.0
numpy>=1.21.0
openstacksdk>=0.61.0